Conflicts:
	README
	fmrd-views.sql
	fmrd.sql

+++ installdb-fmrd.sh
@@ -2,15 +2,20 @@
 
 #    Title: installdb-fmrd.sh
 # Synopsis: Setup FMRD schema and pre-load lookup tables
-#   Format: ./installdb-fmrd.sh
+#   Format: ./installdb-fmrd.sh <database_name>
 #     Date: 2010-08-02
 #  Version: 0.8
 #   Author: Howard Hamilton, Soccermetrics Research & Consulting, LLC
 
-dropdb $1
-createdb $1
+if [ "$#" -eq 1 ]
+then
+    dropdb $1
+    createdb $1
 
-psql -f fmrd.sql $1
-psql -f fmrd-views.sql $1
+    psql -f fmrd.sql $1
+    psql -f fmrd-views.sql $1
 
-./PreloadTables.pl $1
+    ./PreloadTables.pl $1
+else
+    echo "Usage: installdb-fmrd.sh <database_name>"
+fi
